feat: refactor, remove unneeded text
